requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.34.101
botocore==1.34.101
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.6
daphne==4.1.2
defusedxml==0.8.0rc2
Django==4.2.13
django-cors-headers==4.3.1
django-softdelete==0.11.4
django-storages==1.14.3
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jmespath==1.0.1
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.0.8
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.1
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
waitress==3.0.0
whitenoise==6.6.0
zope.interface==7.0.3
